fix: fix default region for aws
